# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Administration = typeof Administration[keyof typeof Administration];


export const Administration = {
  FPS_Bouw: 'FPS Bouw',
  FPS_Brandpreventie: 'FPS Brandpreventie',
  FPS_Onderhoud: 'FPS Onderhoud',
  'FPS_Bouw_&_Renovatie': 'FPS Bouw & Renovatie',
  Futur_Holding: 'Futur Holding',
} as const;

export type DataType = typeof DataType[keyof typeof DataType];


export const DataType = {
  sales_invoice: 'sales_invoice',
  purchase_invoice: 'purchase_invoice',
  bank_transaction: 'bank_transaction',
  timesheet: 'timesheet',
  payroll_journal: 'payroll_journal',
} as const;

export type Provenance = typeof Provenance[keyof typeof Provenance];


export const Provenance = {
  Connect: 'Connect',
  AccountView: 'AccountView',
  manual_import: 'manual_import',
} as const;

export type ArchiveRecordPayload = { [key: string]: unknown };

export interface ArchiveRecord {
  id: number;
  dataType: DataType;
  administration: Administration;
  periodDate: string;
  provenance: Provenance;
  ingestedAt: string;
  archivedAt: string;
  payload: ArchiveRecordPayload;
  /** @nullable */
  documentReference?: string | null;
  /** @nullable */
  storedChecksum?: string | null;
  isInternal?: boolean;
  /** @nullable */
  correctionOfId?: number | null;
  /** @nullable */
  omissionNotes?: string | null;
}

export type CorrectionInputPayload = { [key: string]: unknown };

export interface CorrectionInput {
  dataType: DataType;
  administration: Administration;
  periodDate: string;
  provenance: Provenance;
  payload: CorrectionInputPayload;
  /** @nullable */
  documentReference?: string | null;
  /** @nullable */
  storedChecksum?: string | null;
  isInternal?: boolean;
  omissionNotes: string;
}

export interface ArchiveSummary {
  revenue: number;
  purchases: number;
  hours: number;
  receivables: number;
  payables: number;
  recordCount: number;
  internalInvoiceCount: number;
  completenessPercent: number;
}

export type CompletenessStatusStatus = typeof CompletenessStatusStatus[keyof typeof CompletenessStatusStatus];


export const CompletenessStatusStatus = {
  complete: 'complete',
  partial: 'partial',
  not_loaded: 'not_loaded',
} as const;

export interface CompletenessStatus {
  administration: Administration;
  dataType: DataType;
  status: CompletenessStatusStatus;
  /** @nullable */
  completeFrom: string | null;
  presentFields: string[];
  missingFields: string[];
  recordCount: number;
}

export type DocumentIntegrityStatus = typeof DocumentIntegrityStatus[keyof typeof DocumentIntegrityStatus];


export const DocumentIntegrityStatus = {
  verified: 'verified',
  mismatch: 'mismatch',
  unavailable: 'unavailable',
  not_referenced: 'not_referenced',
} as const;

export interface DocumentIntegrity {
  recordId: number;
  status: DocumentIntegrityStatus;
  /** @nullable */
  documentReference: string | null;
  /** @nullable */
  storedChecksum: string | null;
  /** @nullable */
  actualChecksum?: string | null;
  checkedAt: string;
  message: string;
}

export interface ConsolidationElimination {
  recordId: number;
  invoiceNumber: string;
  fromAdministration: Administration;
  toAdministration: Administration;
  amount: number;
  reason: string;
}

export interface ConsolidationView {
  grossRevenue: number;
  internalEliminations: number;
  consolidatedRevenue: number;
  grossPurchases: number;
  consolidatedPurchases: number;
  eliminations?: ConsolidationElimination[];
}

export type AdministrationParameter = typeof AdministrationParameter[keyof typeof AdministrationParameter];


export const AdministrationParameter = {
  FPS_Bouw: 'FPS Bouw',
  FPS_Brandpreventie: 'FPS Brandpreventie',
  FPS_Onderhoud: 'FPS Onderhoud',
  'FPS_Bouw_&_Renovatie': 'FPS Bouw & Renovatie',
  Futur_Holding: 'Futur Holding',
} as const;

export type PeriodFromParameter = string;

export type PeriodToParameter = string;

export type GetArchiveSummaryParams = {
administration?: AdministrationParameter;
periodFrom?: PeriodFromParameter;
periodTo?: PeriodToParameter;
};

export type GetArchiveCompletenessParams = {
administration?: AdministrationParameter;
};

export type ListArchiveRecordsParams = {
dataType?: ListArchiveRecordsDataType;
administration?: AdministrationParameter;
search?: string;
periodFrom?: PeriodFromParameter;
periodTo?: PeriodToParameter;
/**
 * @minimum 1
 * @maximum 250
 */
limit?: number;
};

export type ListArchiveRecordsDataType = typeof ListArchiveRecordsDataType[keyof typeof ListArchiveRecordsDataType];


export const ListArchiveRecordsDataType = {
  sales_invoice: 'sales_invoice',
  purchase_invoice: 'purchase_invoice',
  bank_transaction: 'bank_transaction',
  timesheet: 'timesheet',
  payroll_journal: 'payroll_journal',
} as const;

export type GetConsolidationParams = {
periodFrom?: PeriodFromParameter;
periodTo?: PeriodToParameter;
};

export type ExportArchiveParams = {
format: ExportArchiveFormat;
dataType?: string;
administration?: AdministrationParameter;
periodFrom?: PeriodFromParameter;
periodTo?: PeriodToParameter;
};

export type ExportArchiveFormat = typeof ExportArchiveFormat[keyof typeof ExportArchiveFormat];


export const ExportArchiveFormat = {
  csv: 'csv',
  json: 'json',
} as const;