# lib/api-zod/src/generated/types/administration.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type Administration = typeof Administration[keyof typeof Administration];


export const Administration = {
  FPS_Bouw: 'FPS Bouw',
  FPS_Brandpreventie: 'FPS Brandpreventie',
  FPS_Onderhoud: 'FPS Onderhoud',
  'FPS_Bouw_&_Renovatie': 'FPS Bouw & Renovatie',
  Futur_Holding: 'Futur Holding',
} as const;