# lib/api-zod/src/generated/types/archiveRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Administration } from './administration';
import type { ArchiveRecordPayload } from './archiveRecordPayload';
import type { DataType } from './dataType';
import type { Provenance } from './provenance';

export interface ArchiveRecord {
  id: number;
  dataType: DataType;
  administration: Administration;
  periodDate: Date;
  provenance: Provenance;
  ingestedAt: Date;
  archivedAt: Date;
  payload: ArchiveRecordPayload;
  /** @nullable */
  documentReference?: string | null;
  /** @nullable */
  storedChecksum?: string | null;
  isInternal?: boolean;
  /** @nullable */
  correctionOfId?: number | null;
  /** @nullable */
  omissionNotes?: string | null;
}