# lib/api-zod/src/generated/types/archiveSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ArchiveSummary {
  revenue: number;
  purchases: number;
  hours: number;
  receivables: number;
  payables: number;
  recordCount: number;
  internalInvoiceCount: number;
  completenessPercent: number;
}