# lib/api-zod/src/generated/types/completenessStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Administration } from './administration';
import type { CompletenessStatusStatus } from './completenessStatusStatus';
import type { DataType } from './dataType';

export interface CompletenessStatus {
  administration: Administration;
  dataType: DataType;
  status: CompletenessStatusStatus;
  /** @nullable */
  completeFrom: Date | null;
  presentFields: string[];
  missingFields: string[];
  recordCount: number;
}