# lib/api-zod/src/generated/types/consolidationElimination.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Administration } from './administration';

export interface ConsolidationElimination {
  recordId: number;
  invoiceNumber: string;
  fromAdministration: Administration;
  toAdministration: Administration;
  amount: number;
  reason: string;
}