# lib/api-zod/src/generated/types/provenance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type Provenance = typeof Provenance[keyof typeof Provenance];


export const Provenance = {
  Connect: 'Connect',
  AccountView: 'AccountView',
  manual_import: 'manual_import',
} as const;